XPath: widen the string-value cache bucket index to size_t
Audit follow-up. The per-evaluate string-value cache stored its open-addressing
bucket payload (entry index + 1) as uint32_t, the same size_t->uint32_t
truncation class just fixed for the document-order ordinal. count is bounded by
distinct nodes touched in one evaluate and the entries[] store (each with an
owned string copy) OOMs long before 2^32, so it is not reachable in practice —
but a truncated index would silently return another node's cached string-value,
which the engine must never do. Widen buckets to size_t so the index can't
truncate; the cast at the insert is dropped. (The SplitMix bucket hash keeps its
intentional uint32_t output.)

The rest of the audit found no reachable gap: the remaining narrow-int / +1 /
*2 spots (translate's from.len+1, source_loc offset+1, the reindex-trigger
multiplications, and the (long) casts of single Lexbor object lengths) are each
bounded by max_string_bytes, the node-set cap, or physical memory (a single
in-memory object cannot exceed LONG_MAX), and fail closed by OOM before any
overflow.

No behaviour change on reachable inputs. Verified: clean compile (no warnings),
C self-test true, security:clint 0 findings, spec 368/0/4 pending, ASan/UBSan
clean, fuzz:sanitize 0 findings, Nokogiri XPath differential 0 divergence /
0 order-only diff.

+++ ext/makiri/xpath/mkr_xpath_internal.h
@@ -480,8 +480,9 @@ typedef struct {
   /* Open-addressing index: node pointer -> (entry index + 1); 0 = empty slot.
    * Keeps lookup/insert O(1) so per-node predicate comparisons over a large
    * node-set don't degrade to O(n^2). entries[] is the ordered store (so
-   * truncate to a snapshot stays cheap); buckets is just an index into it. */
-  uint32_t *buckets;
+   * truncate to a snapshot stays cheap); buckets is just an index into it.
+   * size_t (not uint32_t) so the stored entry index can never truncate. */
+  size_t   *buckets;
   size_t    bucket_cap; /* power of two, or 0 */
 } mkr_str_cache_t;
